+++ changelog.md
@@ -2,7 +2,32 @@
 
 ## dev
 
-## 0.5.3 (master)
+Table Widget:
+- Added support `options.filterValue`..
+- Fixed filtering when row data contains HTML tags.
+
+Added `LX.stripHTML(html)` to get clean text from HTML string.
+Fixed all "popup" menus position in scrolled window.
+Added support for options `side`, `active` and `offset` in `LX.asTooltip(trigger, content, options)`.
+Allowed users to register extra icons in LX.ICONS using `LX.registerIcon`;
+Icons and Customization docs updated.
+
+## 0.5.4 (master)
+
+Widgets:
+- Added support for disabling resize in TextArea Widget `options.resize: false`.
+- Added support for using LX.ICONS on Button Widgets.
+- Tooltips supported in Button Widget `options.tooltip`. Uses either `options.title` or `name` as content.
+
+`LX.makeContainer` now accepts `innerHTML` and `parent` params.
+Started work on tooltips `LX.asTooltip(trigger, content)`.
+Added support for adding a class to LexGUI root on init `options.rootClass`.
+Added support for adding a class to Sidebar/Menubar areas on `Area.addSidebar` and `Area.addMenubar`.
+Added `options.asElement` to return `LX.badge` as HTMLElement instead of html string.
+Support for using LX.ICONS on sidebar entry icons. Keep FA icons as legacy by now.
+Minor CSS improvements. Added more class utilities.
+
+## 0.5.3
 
 Widgets:
 - Fixed Select Widget filter issues.